backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.4
python-multipart>=0.0.9
psycopg2-binary>=2.9.7
google-api-python-client>=2.110.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
anyio>=4.5
pydantic-settings>=2.5.2
asyncpg>=0.28.0
supabase>=2.22.0
requests>=2.31.0
dedalus-labs
agentmail
resend
langgraph>=1.0.0
langchain-core>=1.0.0
sentry-sdk[fastapi]>=2.0.0
